Ability to play predefined animations.
Added the ability to play pre-defined animations as either a loop or a
one-shot animation. Callers can now pass a `:frames` option to the
`%Xebow.Animation{}` struct, which is a list of `Xebow.AnimationFrames`.

Additionally, for "dynamic" animations, I added the ability to track the
last few frames of animation. I think that we need this information in
order to do some advanced animations like "wave" effects that reflect
off of the edges of the matrix and other key presses.