Remove use of GPIO.set_interrupt/3
Using the built-in `GPIO.set_interrupt/3` function was not so flexible
because we can't tune the `:supress_glitches` parameter.

We ripped out the use of `GPIO.set_interrupt/3` and replaced it with a
custom polling mechanism that reads the pin values and fires off AFK
state updates.

Additionally, we were able to remove a lot of cruft surrounding the
initialization of the Keys `GenServer`.

Co-authored-by: Alex McLain <alex@alexmclain.com>